# api/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task API",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/health/live": {
      "get": {
        "summary": "Liveness (bonus)",
        "responses": {
          "200": { "description": "Process up" }
        }
      }
    },
    "/health/ready": {
      "get": {
        "summary": "Readiness — DB ping (bonus)",
        "responses": {
          "200": { "description": "Postgres reachable" },
          "503": { "description": "Not ready" }
        }
      }
    },
    "/status/{requestId}": {
      "get": {
        "summary": "Request lifecycle from DB (bonus)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "States + audits for this request_id" },
          "400": { "description": "Bad UUID" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "No rows for id" }
        }
      }
    },
    "/dashboard": {
      "get": {
        "summary": "Simple HTML dashboard over recent audit/state rows (bonus)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "text/html" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/auth/token": {
      "post": {
        "summary": "Issue Token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string", "example": "demo" },
                  "password": { "type": "string", "example": "demo" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "JWT issued" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/task": {
      "post": {
        "summary": "Create Task",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "payload": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              },
              "example": { "payload": { "hello": "world" } }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task queued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["requestId", "handledBy"],
                  "properties": {
                    "requestId": { "type": "string", "format": "uuid" },
                    "handledBy": { "type": "string", "description": "API instance (SERVICE_NAME), e.g. api1" },
                    "status": { "type": "string", "example": "queued" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "502": { "description": "Broker error" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}